Separated wasm-rpc specific part of dynamic linking
