+++ golem-worker-executor-base/src/services/rpc.rs
@@ -155,7 +155,9 @@ impl From<golem_wasm_rpc::RpcError> for RpcError {
             golem_wasm_rpc::RpcError::ProtocolError(details) => Self::ProtocolError { details },
             golem_wasm_rpc::RpcError::Denied(details) => Self::Denied { details },
             golem_wasm_rpc::RpcError::NotFound(details) => Self::NotFound { details },
-            golem_wasm_rpc::RpcError::RemoteInternalError(details) => Self::RemoteInternalError { details },
+            golem_wasm_rpc::RpcError::RemoteInternalError(details) => {
+                Self::RemoteInternalError { details }
+            }
         }
     }
 }
